fc: fix motor control sign error

+++ src/core/controller.h
@@ -19,7 +19,7 @@ void attitude_pd_control(pid_control_t *pid, float ahrs_attitude,
 void yaw_rate_p_control(pid_control_t *pid, float setpoint_yaw_rate,
 			float angular_velocity);
 
-void motor_control(volatile float throttle_percentage, uint16_t roll_ctrl_precentage,
-		   uint16_t pitch_ctrl_precentage, float yaw_ctrl_precentage);
+void motor_control(volatile float throttle_percentage, float roll_ctrl_precentage,
+		   float pitch_ctrl_precentage, float yaw_ctrl_precentage);
 
 #endif
